+++ lib/api/src/index.tsx
@@ -31,11 +31,18 @@ import initStories, { SubState as StoriesSubState, SubAPI as StoriesAPI } from '
 import initRefs, {
   SubState as RefsSubState,
   SubAPI as RefsAPI,
-  defaultMapper,
   getSourceType,
-  Mapper,
 } from './modules/refs';
-import { StoriesRaw } from './lib/stories';
+import {
+  StoriesRaw,
+  StoriesHash,
+  Story,
+  Root,
+  Group,
+  isGroup,
+  isRoot,
+  isStory,
+} from './lib/stories';
 import initLayout, {
   ActiveTabs,
   SubState as LayoutSubState,
@@ -370,7 +377,17 @@ export function useStorybookApi(): API {
   return api;
 }
 
-export { ManagerConsumer as Consumer, ManagerProvider as Provider };
+export {
+  ManagerConsumer as Consumer,
+  ManagerProvider as Provider,
+  StoriesHash,
+  Story,
+  Root,
+  Group,
+  isGroup,
+  isRoot,
+  isStory,
+};
 
 export interface EventMap {
   [eventId: string]: Listener;